# By Aleksei Kanash
# Via Aleksei Kanash
* commit '8713e351f9afe52fd4f0394b45e01eafbaba5cb8':
  [DF-6727] Fix merge.
  [DF-6727] removed thumb from box2d and review fix.
  [DF-6727] Fix warnings and disabled mno-thumb for iOS.

+++ Sources/Internal/Autotesting/AutotestingDB.cpp
@@ -198,7 +198,7 @@ namespace DAVA
 		va_start(vl, text);
 		char tmp[4096] = { 0 };
 		vsnprintf(tmp, sizeof(tmp) - 2, text, vl);
-		file->Write(text, sizeof(char) *strlen(tmp));
+		file->Write(text, static_cast<uint32>(sizeof(char)*strlen(tmp)));
 		file->Release();
 		va_end(vl);
 	}
